services/alarm_broker/pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alarm-broker"
version = "0.1.0"
description = "Alarm Broker (MVP) – Yealink → Broker → Zammad/SMS/Signal with ACK and escalation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "python-multipart>=0.0.12",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "sqlalchemy>=2.0",
  "greenlet>=3.0",
  "alembic>=1.13",
  "asyncpg>=0.29",
  "psycopg[binary]>=3.2",
  "arq>=0.26",
  "httpx>=0.27",
  "tenacity>=8.5",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8",
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "respx>=0.21",
  "aiosqlite>=0.20",
  "bandit>=1.7",
  "pip-audit>=2.7",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["alarm_broker*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
